backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.1.0
django-filter==23.2
pytz==2023.3